[mysql]: fix the normalizedSplitRecords logic
When nomarlize the snapshotRecords and binlogRecords, we should
remove from snapshotRecords by binlogRecords.keys().

Short version reason: it's the rule from DBLog paper

Long version reason. TLDR;

Let't see the followling illustration:

```
    snapshot readview(sr)
            |
            v
------lw---------hw-----    <- binlog stream
```

For RR isolation level, the readview is a point corresponding to
the binlog stream, let's name it sr.

1. For an update binlog(ub):

If lw <= ub <= sr, the normalizedBinlogRecord first add the
new value(i.e the updateAfter), and then add the snapshotRecords,
so there will be two record for this key, both contain the new value,
there is no problem since the downstream use the upsert logic.

But if sr < ub <= hw, the snapshotRecords contains the old
value, and the normalizedBinlogRecord first add the new value,
and then add the old value from the snapshotRecords, now there
is problem, since the upsert logic will apply the old value
to the downstream, data inconsistency ;(

2. For an delete binlog(db)

If lw <= db <= sr, the snapshotRecords does not contain the
corresponding key, and if sr < db <= hw, the snapshotRecords
contains the key, so the logic `if contain then delete` is
just ok, no need to throw Exception if snapshotRecords does
not contains the key.

3. For an create binlog(cb)

If lw <= db <= sr, there will be two create binlog in the
normalizedBinlogRecords with the same value, no problem
since downstream apply the upsert logic.

So, a simple solution to update and delete problem is to
delete the corresponding key from snapshotRecords.
For create, delete from snapshotRecords will eliminate
the duplication.

For RC isolation level, delete from snapshotRecords first
will also work, I will leave the analysis for those who
will see this patch and interested.

Signed-off-by: 元组 <zhaojunwang.zjw@alibaba-inc.com>