[mysql] expose Enumerator metrics via Source Event
I opened a PR #707 last week, which hope to expose some
Enumerator metrics directly to `SplitEnumeratorContext.metricGroup()`,
but unfortunately, it return null for the `MetricGroup`. After
some investigation of FLIP-27, it says `The SourceCoordinators
will run in the JobMaster, but not as part of the ExecutionGraph`.

So there is noway to expose the Enumerator metrics ;(

In this PR, I use the RPC Gateway, a.k.a SourceEvent to transfer
the Enumerator metrics to each SourceReader.

I basicly added 5 metrics:

- alreadyProcessedTableCount  // Already processed table count for this flink job.
- remainingTableCount	// Remaining table count for this flink job to handle.
- finishedSplitCount	// Processed splits count for this flink job.
- assignedSplitCount	// Assigned but not finished splits count for this flink job.
- remainingSplitCount	// Remaining split count for the currently processing table.

For now I just use the `FinishedSnapshotSplitsAckEvent` to
carry the above metrics, we may alternatively send a
request event to query the metrics in `MySqlSourceReader#start`
and `MySqlSourceReader#onSplitFinished`, that might need
some future discussion.

Signed-off-by: 元组 <zhaojunwang.zjw@alibaba-inc.com>