[mysql] optimize scanNewlyAddedTable
In current implemtation, `MySqlSourceEnumerator` send suspend event to
`MySqlSourceReader` in `start` and `syncWithReader` periodly, there is
a large chance the event sent from `start` be lost if the
`MySqlSourceReader` has not been registered.

In this PR, instead of waiting for the suspend event, `MySqlSourceReader`
suspend the binglog split **eagerly** by checking the `sourceConfig` if
there is newly added table, this simplified the `MySqlSourceReader` logic,
the `MySqlSplitReader` need not to be told to stop.

We use the flink stop savepoint to take care of the binlogSplit state.

Signed-off-by: 元组 <zhaojunwang.zjw@alibaba-inc.com>