[mysql] accelerate suspend speed
`suspendBinlogReaderIfNeed` broadcasts `SuspendBinlogReaderEvent`
when splitAssigner is in suspend status, but there are chances
that this happens before `SourceReader` registered to
`SourceEnumerator`, in that case, we will need to wait
`CHECK_EVENT_INTERVAL` a.k.a 30s to resend the event message,
which can be annoying.

In this patch, simply send `SuspendBinlogReaderEvent` message
when `SourceReader` registers if the assigner's status is
suspended.

Signed-off-by: 元组 <zhaojunwang.zjw@alibaba-inc.com>